+++ rules/src/firestore.test.ts
@@ -17,9 +17,9 @@
 // import { itAllowsReadForUser } from "./util/tests/reading";
 
 import { clearFirestore } from "./util";
-import { itDeniesByDefault } from "./util/common/denies-by-default";
+import { itDeniesByDefault } from "./util/access/denies-by-default";
 
-const dashboard = "dashboard";
+// const dashboard = "dashboard";
 
 // const tokenDocPath = () => `tokens/${v4()}`;
