+++ rules/src/util/access/common/allows-delete-for-user.ts
@@ -1,0 +1,9 @@
+import { setupDoc } from "../..";
+
+export function itAllowsDeleteForUser(document: string, uid: string) {
+  it(`Allows '${uid}' to delete '${document}'`, async () => {
+    const ref = await setupDoc(document, uid);
+
+    await expect(ref.delete()).toAllow();
+  });
+}
