+++ rules/src/util/access/common/denies-delete-for-user.ts
@@ -1,0 +1,9 @@
+import { setupDoc } from "../..";
+
+export function itDeniesDeleteForUser(document: string, uid: string) {
+  it(`Denies '${uid}' to delete '${document}'`, async () => {
+    const ref = await setupDoc(document, uid);
+
+    await expect(ref.delete()).toDeny();
+  });
+}
