+++ rules/src/util/access/common/index.ts
@@ -1,0 +1,8 @@
+export * from "./allows-create-for-user";
+export * from "./allows-delete-for-user";
+export * from "./allows-read-for-user";
+export * from "./allows-update-for-user";
+export * from "./denies-create-for-user";
+export * from "./denies-delete-for-user";
+export * from "./denies-read-for-user";
+export * from "./denies-update-for-user";
